ollama: split raw /api/generate into its own endpoint
Also added a lot of TypeAliases to make things easier to reason about;
should really invest into Pydantic models, if this spirals in
complexity. This should be the end of it, though.

Additionally, standardized naming schema for a few things:

- TemplatedPromptText is just tokens, with LLM-specific templating
  applied.
- Also removed the use of "raw", since it's inconsistent and only
  matters to the Ollama API options, anyway.
- TODO: RequestInterceptor handles bytes and JSON and str everywhere,
  but I don't think that'll get any clearer without Pydantic.

TODO: Need to start logging client-side requests/responses, for future
inspection. Which is easier than writing brittle code, but not great for
potential end users.