xcode wip: Start working on better ChatSequence View
Ideally we'd have an iOS TabView for the text entry at the bottom, since
we'll need to support documents and model-shifting.

Next thing to do should probably be consistent InferenceModel
"identifiers"; closely related to EmbeddingModels, so I'm not sure how
much we'll want to split the class hierarchy for that.