xcode: add a bunch of not-working settings
Added a bunch of Pickers and Toggles; this feels like it would be a huge
mess to work with, and we should just have something much more visually
distinguishable. It's hard to do this in the limited space of the
message box, because we would probably want explanatory diagrams or make
immediate changes visible (submit/submitWithRetrieval buttons).

Or I guess there could be a bunch of context menus, too. Right click the
submit button, right click the entry field. What to do with the gnarlier
options, though? An OSV-level settings overlay?