SequenceViewModel: don't submit if still receiving
This happens because `.keyboardShortcut` is still enabled even when the
button is disabled, so we can get multiple submits at the same time.

This handles double-submits correctly, by doing nothing in the second
case.

However, because of how the submit button transforms into a stop button,
cmd-enter will also stop receiving. Which is non-intuitive, but the
cmd-enter shortcut isn't surfaced to the user, either.