OSV: tidy up local ChatSequence construction
The key was filling in all the missing fields in BlankSequenceViewModel.
Because there's a lot of logic dependent on things like parentSequences,
I guess.

There's also a bit of work around `DispatchQueue.main.sync`, since we're
relying on order of work to happen. So, we have to organize it so it
doesn't really matter when the work happens. Which is an
outdated/backwards model for Swift concurrency.