OMV: add button to hide reasoning for R1-like models
To detect reasoning, we have the client check for .contains("</think>").

- This means OneMessageView has to hold two MessageLikes, one with the
  reasoning and one without, so we can toggle between them.
- This needs us to replace the markdown render cache; implementation-wise
  it was only operating on strings _anyway_, but an alternative would be
  MessageLike + a Bool.

This also lets us hide the reasoning for an in-progress
TemporaryChatMessage; the UI for this becomes terrible (a bunch of text
disappears), but since we have no way of marking or recognizing a model
as reasoning, we're just making guesses _anyway_.

NB The way reasoning content is delimited is limited by the
llama.cpp/ollama API, and implementing a server-side solution is likely
a non-starter, until we get something more stable to work with.

Also note that follow-on ChatMessages will have the server parse the
entire reasoning chain, since it's only hidden on the client side, and
not removed from the chat history.