Settings wip: Change animateNewResponseText from a bool to an enum
This way, user can decide to animate, even for markdown messages.

This will result in strang cross-fading artifacts, since behavior
depends on the markdown rendering library we're using.

NB This isn't working yet, because we're not doing the check correctly.
So far, just the Settings plumbing matters (we've created a new
`defaults` entry, thus the rename).